+++ R/theme_inter.r
@@ -1,39 +1,51 @@
-#' ggplot2 theme using custom font
-#' @rdname theme_font
-#' @param ... Arguments passed to `plummy::theme_custom()`
+#' @rdname import_font
+#' @importFrom sysfonts font_add
 #' @export
 #'
-theme_inter <- function(...) {
-    theme_custom(
-        base_family = "Inter",
-        base_size = 10,
-        plot_title_family = "Inter Semi Bold",
-        plot_title_margin = 8,
-        subtitle_family = "Inter",
-        subtitle_margin = 13,
-        strip_text_family = "Inter Medium",
-        caption_family = "Inter Light",
-        axis_title_family = "Inter",
-        axis_text_family = "Inter Light",
-        ...
-    )
+import_inter <- function() {
+  font_dir <- system.file("fonts", "inter", package = "plummy")
+
+  font_add(
+    family = "inter",
+    regular = file.path(font_dir, "Inter-Regular.ttf"),
+    bold = file.path(font_dir, "Inter-SemiBold.ttf"),
+    italic = file.path(font_dir, "Inter-Italic.ttf"),
+    bolditalic = file.path(font_dir, "Inter-SemiBoldItalic.ttf")
+  )
+
+  # font_add(family = "inter-100-thin", file.path(font_dir, "Inter-Thin.ttf"))
+  # font_add(family = "inter-100-thinitalic", file.path(font_dir, "Inter-ThinItalic.ttf"))
+  # font_add(family = "inter-200-extralight", file.path(font_dir, "Inter-ExtraLight.ttf"))
+  # font_add(family = "inter-200-extralightitalic", file.path(font_dir, "Inter-ExtraLightItalic.ttf"))
+  font_add(family = "inter-300-light", file.path(font_dir, "Inter-Light.ttf"))
+  font_add(family = "inter-300-lightitalic", file.path(font_dir, "Inter-LightItalic.ttf"))
+  font_add(family = "inter-400-regular", file.path(font_dir, "Inter-Regular.ttf"))
+  font_add(family = "inter-400-regularitalic", file.path(font_dir, "Inter-Italic.ttf"))
+  font_add(family = "inter-500-medium", file.path(font_dir, "Inter-Medium.ttf"))
+  font_add(family = "inter-500-mediumitalic", file.path(font_dir, "Inter-MediumItalic.ttf"))
+  font_add(family = "inter-600-semibold", file.path(font_dir, "Inter-SemiBold.ttf"))
+  font_add(family = "inter-600-semibolditalic", file.path(font_dir, "Inter-SemiBoldItalic.ttf"))
+  font_add(family = "inter-700-bold", file.path(font_dir, "Inter-Bold.ttf"))
+  # font_add(family = "inter-700-bolditalic", file.path(font_dir, "Inter-BoldItalic.ttf"))
+  # font_add(family = "inter-800-extrabold", file.path(font_dir, "Inter-ExtraBold.ttf"))
+  # font_add(family = "inter-800-extrabolditalic", file.path(font_dir, "Inter-ExtraBoldItalic.ttf"))
+  # font_add(family = "inter-900-black", file.path(font_dir, "Inter-Black.ttf"))
 }
 
-#' Import font for use in charts
-#'
-#' @rdname import_font
-#' @importFrom extrafont font_import
+#' @rdname theme_grid
+#' @param ... Arguments passed to `plummy::theme_grid()`
 #' @export
-#' 
-import_inter <- function() {
-    font_dir <- system.file("fonts", "inter", package = "plummy")
-
-    suppressWarnings(suppressMessages(extrafont::font_import(font_dir, prompt = FALSE)))
-
-    message(
-        sprintf(
-            "You will likely need to install these fonts on your system as well.\n\nYou can find them in [%s]",
-            font_dir
-        )
-    )
+#'
+theme_grid_inter <- function(...) {
+  theme_grid(
+    base_family = "inter-400-regular",
+    base_size = 10,
+    title_family = "inter-600-semibold",
+    subtitle_family = "inter-400-regular",
+    strip_text_family = "inter-500-medium",
+    caption_family = "inter-300-light",
+    axis_title_family = "inter-400-regular",
+    axis_text_family = "inter-300-light",
+    ...
+  )
 }