+++ NEWS.md
@@ -1,3 +1,14 @@
+# plummy 0.2
+
+- 01-Sep-2024
+- Restructured code
+- Switched extrafont to showtext and sysfonts
+- Added more font variants
+- Added Lato font
+- Created wrapper function plummy()
+- Changed theme_custom() to theme_grid()
+- Added palettes: morris, okabeito, strong
+
 # plummy 0.1
 
 - 05-May-2023